Move asset code into separate file
The goaci.go already looks like a kitchen sink with two year old
leftovers. Let's clean it up a bit.
