+++ goaci.go
@@ -6,7 +6,6 @@ import (
 	"compress/gzip"
 	"flag"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
 	"os/exec"
@@ -108,60 +107,6 @@ func getOptions() *options {
 	}
 
 	return opts
-}
-
-func copyTree(src, dest string) error {
-	return filepath.Walk(src, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-		rootLess := path[len(src):]
-		target := filepath.Join(dest, rootLess)
-		mode := info.Mode()
-		switch {
-		case mode.IsDir():
-			err := os.Mkdir(target, mode.Perm())
-			if err != nil {
-				return err
-			}
-		case mode.IsRegular():
-			srcFile, err := os.Open(path)
-			if err != nil {
-				return err
-			}
-			destFile, err := os.Create(target)
-			if err != nil {
-				return err
-			}
-			if _, err := io.Copy(destFile, srcFile); err != nil {
-				return err
-			}
-		case mode&os.ModeSymlink == os.ModeSymlink:
-			symTarget, err := filepath.EvalSymlinks(path)
-			if err != nil {
-				return err
-			}
-			if !filepath.IsAbs(symTarget) {
-				dirPart := filepath.Dir(path)
-				testPath := filepath.Join(dirPart, symTarget)
-				symTarget = filepath.Clean(testPath)
-			}
-			if strings.HasPrefix(symTarget, src) {
-				relTarget, err := filepath.Rel(filepath.Dir(path), symTarget)
-				if err != nil {
-					return err
-				}
-				if err := os.Symlink(relTarget, target); err != nil {
-					return err
-				}
-			} else {
-				return fmt.Errorf("Symlink %s points to %s, which is outside asset %s", path, symTarget, src)
-			}
-		default:
-			return fmt.Errorf("Unsupported node (%s) in assets, only regular files, directories and symlinks pointing to node inside asset are supported.", path, mode.String())
-		}
-		return nil
-	})
 }
 
 func main() {
@@ -319,38 +264,8 @@ func main() {
 	}
 	debug("moved binary to:", ep)
 
-	// Prepare assets
-	for _, asset := range opts.assets {
-		splitAsset := filepath.SplitList(asset)
-		if len(splitAsset) != 2 {
-			die("Malformed asset option: '%v' - expected two absolute paths separated with %v", asset, listSeparator())
-		}
-		ACIAsset := splitAsset[0]
-		localAsset := splitAsset[1]
-		if !filepath.IsAbs(ACIAsset) {
-			die("Malformed asset option: '%v' - ACI asset has to be absolute path", asset)
-		}
-		if !filepath.IsAbs(localAsset) {
-			die("Malformed asset option: '%v' - local asset has to be absolute path", asset)
-		}
-		fi, err := os.Stat(localAsset)
-		if err != nil {
-			die("Error stating %v: %v", localAsset, err)
-		}
-		if fi.Mode().IsDir() || fi.Mode().IsRegular() {
-			ACIBase := filepath.Base(ACIAsset)
-			ACIAssetSubPath := filepath.Join(rfs, filepath.Dir(ACIAsset))
-			err := os.MkdirAll(ACIAssetSubPath, 0755)
-			if err != nil {
-				die("Failed to create directory tree for asset '%v': %v", asset, err)
-			}
-			err = copyTree(localAsset, filepath.Join(ACIAssetSubPath, ACIBase))
-			if err != nil {
-				die("Failed to copy assets for '%v': %v", asset, err)
-			}
-		} else {
-			die("Can't handle %v - not a file, not a dir", fi.Name())
-		}
+	if err := PrepareAssets(opts.assets, rfs); err != nil {
+		die("%v", err)
 	}
 
 	exec := []string{filepath.Join("/", fn)}
